Organize tests and fix dependent tests
