+++ catalog/tests/dags/common/storage/test_audio.py
@@ -61,7 +61,7 @@ def test_AudioStore_add_item_adds_realistic_audio_to_buffer():
     audio_store.add_item(
         foreign_identifier="01",
         foreign_landing_url="https://audios.org/audio01",
-        url="https://audios.org/audio01.jpg",
+        url="https://audios.org/audio01.mp3",
         license_info=license_info,
         ingestion_type="provider_api",
     )
@@ -73,25 +73,25 @@ def test_AudioStore_add_item_adds_multiple_audios_to_buffer():
     audio_store.add_item(
         foreign_identifier="01",
         foreign_landing_url="https://audios.org/audio01",
-        url="https://audios.org/audio01.jpg",
+        url="https://audios.org/audio01.mp3",
         license_info=PD_LICENSE_INFO,
     )
     audio_store.add_item(
         foreign_identifier="02",
         foreign_landing_url="https://audios.org/audio02",
-        url="https://audios.org/audio02.jpg",
+        url="https://audios.org/audio02.mp3",
         license_info=PD_LICENSE_INFO,
     )
     audio_store.add_item(
         foreign_identifier="03",
         foreign_landing_url="https://audios.org/audio03",
-        url="https://audios.org/audio03.jpg",
+        url="https://audios.org/audio03.mp3",
         license_info=PD_LICENSE_INFO,
     )
     audio_store.add_item(
         foreign_identifier="04",
         foreign_landing_url="https://audios.org/audio04",
-        url="https://audios.org/audio04.jpg",
+        url="https://audios.org/audio04.mp3",
         license_info=PD_LICENSE_INFO,
     )
     assert len(audio_store._media_buffer) == 4
@@ -105,25 +105,25 @@ def test_AudioStore_add_item_flushes_buffer(tmpdir):
     audio_store.add_item(
         foreign_identifier="01",
         foreign_landing_url="https://audios.org/audio01",
-        url="https://audios.org/audio01.jpg",
+        url="https://audios.org/audio01.mp3",
         license_info=PD_LICENSE_INFO,
     )
     audio_store.add_item(
         foreign_identifier="02",
         foreign_landing_url="https://audios.org/audio02",
-        url="https://audios.org/audio02.jpg",
+        url="https://audios.org/audio02.mp3",
         license_info=PD_LICENSE_INFO,
     )
     audio_store.add_item(
         foreign_identifier="03",
         foreign_landing_url="https://audios.org/audio03",
-        url="https://audios.org/audio03.jpg",
+        url="https://audios.org/audio03.mp3",
         license_info=PD_LICENSE_INFO,
     )
     audio_store.add_item(
         foreign_identifier="04",
         foreign_landing_url="https://audios.org/audio04",
-        url="https://audios.org/audio04.jpg",
+        url="https://audios.org/audio04.mp3",
         license_info=PD_LICENSE_INFO,
     )
     assert len(audio_store._media_buffer) == 1
@@ -142,22 +142,43 @@ def test_AudioStore_produces_correct_total_audios():
     audio_store.add_item(
         foreign_identifier="01",
         foreign_landing_url="https://audios.org/audio01",
-        url="https://audios.org/audio01.jpg",
+        url="https://audios.org/audio01.mp3",
         license_info=PD_LICENSE_INFO,
     )
     audio_store.add_item(
         foreign_identifier="02",
         foreign_landing_url="https://audios.org/audio02",
-        url="https://audios.org/audio02.jpg",
+        url="https://audios.org/audio02.mp3",
         license_info=PD_LICENSE_INFO,
     )
     audio_store.add_item(
         foreign_identifier="03",
         foreign_landing_url="https://audios.org/audio03",
-        url="https://audios.org/audio03.jpg",
+        url="https://audios.org/audio03.mp3",
         license_info=PD_LICENSE_INFO,
     )
     assert audio_store.total_items == 3
+
+
+@pytest.mark.parametrize(
+    "filetype, url, expected_filetype",
+    [
+        # The value provided prevails over the url extension
+        ("ogg", "http://example.com/audio", "ogg"),
+        ("ogg", "http://example.com/audio.wav", "ogg"),
+        # The filetype is guessed from the URL extension
+        (None, "http://example.com/audio.mp3", "mp3"),
+        (None, "http://example.com/audio.WAV", "wav"),
+        (None, "http://example.com/audio.mid", "mid"),
+        # Unifies filetypes
+        ("midi", "http://example.com/audio.mid", "mid"),
+        (None, "http://example.com/audio.midi", "mid"),
+    ],
+)
+def test_AudioStore_validate_filetype(filetype, url, expected_filetype):
+    audio_store = audio.MockAudioStore("test_provider")
+    actual_filetype = audio_store._validate_filetype(filetype, url)
+    assert actual_filetype == expected_filetype
 
 
 @test_media.INT_MAX_PARAMETERIZATION
